requirements.txt
absl-py==2.2.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
atoti==0.9.4
atoti-client==0.9.4
atoti-core==0.9.4
atoti-server==0.9.4
attrs==25.1.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
et_xmlfile==2.0.0
etl==0.0.1
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.56.0
frozendict==2.4.6
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.3
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.32.0
jax==0.4.30
jaxlib==0.4.30
jdk4py==21.0.4.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaggle==1.7.4.5
keras==2.12.0
kiwisolver==1.4.8
kombu==5.5.3
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
MiniSom==2.3.3
mistune==3.1.3
ml_dtypes==0.5.1
multitasking==0.0.11
namex==0.0.8
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.3.0
oauthlib==3.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.1
pmdarima==2.0.3
prometheus_client==0.22.0
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
pywin32==308
pyzmq==26.2.1
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.1.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
statsmodels==0.13.5
streamlit==1.45.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
ucimlrepo==0.0.7
urllib3==2.3.0
uv==0.6.3
vine==5.1.0
vmdpy==0.2
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.14.1
xgboost==3.0.0
yfinance==0.2.55
